# Conflicts:
#	src/java/br/unesp/rc/habilidades/commands/CriarMembro.java
#	src/java/br/unesp/rc/habilidades/commands/DoEditarMembro.java
#	src/java/br/unesp/rc/habilidades/servlet/CommandHelper.java

+++ src/java/br/unesp/rc/habilidades/commands/DoExcluirMembro.java
@@ -5,8 +5,10 @@
  */
 package br.unesp.rc.habilidades.commands;
 
+import br.unesp.rc.habilidades.beans.Permissao;
 import br.unesp.rc.habilidades.dao.MembroDAO;
 import br.unesp.rc.habilidades.dao.MembroDAOImpl;
+import br.unesp.rc.habilidades.util.PermissaoUtils;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
@@ -18,14 +20,17 @@ public class DoExcluirMembro implements ICommand {
 
     @Override
     public CommandResult execute(HttpServletRequest request, HttpServletResponse response) {
-        System.out.println("HELLO WORLD");
-        int id = Integer.parseInt(request.getParameter("idMembro"));
         
-        System.out.println("HELLO WORLD" + id);
+        if (!PermissaoUtils.hasPermissao(request, Permissao.GERENCIAR_MEMBROS))
+        {
+             return new CommandResult("forbidden");
+        }        
+        
+        int id = Integer.parseInt(request.getParameter("idMembro"));        
+       
         MembroDAO membroDao = new MembroDAOImpl();
         membroDao.remove(id);
-        System.out.println("HELLO WORLD");     
-        
+                
         request.setAttribute("msg_tipo", "alert-success");
         request.setAttribute("msg", "Membro excluído com sucesso");
         request.setAttribute("menu", "adminmembro");
